jwt long ->Integer 타입변경

+++ src/main/java/kr/co/anabada/jwt/JwtUtil.java
@@ -31,28 +31,24 @@ public class JwtUtil {
         this.refreshTokenExpiration = refreshTokenExpiration;
     }
 
-    //Access Token userId만
     public String generateAccessToken(String userId) {
         return generateToken(userId, null, null, null, accessTokenExpiration);
     }
-    //Access Token
-    public String generateAccessToken(String userId, Long userNo, String userType, String nickname) {
+
+    public String generateAccessToken(String userId, Integer userNo, String userType, String nickname) {
         return generateToken(userId, userNo, userType, nickname, accessTokenExpiration);
     }
 
-    //Refresh Token
     public String generateRefreshToken(String userId) {
         return generateToken(userId, null, null, null, refreshTokenExpiration);
     }
 
-    //공통 토큰
-    private String generateToken(String userId, Long userNo, String userType, String nickname, long expiration) {
+    private String generateToken(String userId, Integer userNo, String userType, String nickname, long expiration) {
         var builder = Jwts.builder()
                 .setSubject(userId)
                 .setIssuedAt(new Date())
                 .setExpiration(new Date(System.currentTimeMillis() + expiration));
 
-        // 커스텀 claim
         if (userNo != null) builder.claim("userNo", userNo);
         if (userType != null) builder.claim("userType", userType);
         if (nickname != null) builder.claim("nickname", nickname);
@@ -61,15 +57,11 @@ public class JwtUtil {
                 .signWith(key, SignatureAlgorithm.HS256)
                 .compact();
     }
-    
 
-    
-    //AccessToken 추출
     public String extractAccessToken(HttpServletRequest request) {
         return extractToken(request);
     }
 
-    //내부 공통 토큰 추출
     public String extractToken(HttpServletRequest request) {
         String authHeader = request.getHeader("Authorization");
         if (authHeader != null && authHeader.startsWith("Bearer ")) {
@@ -78,7 +70,6 @@ public class JwtUtil {
         return null;
     }
 
-    //토큰 유효성 검증
     public boolean validateToken(String token) {
         try {
             Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
@@ -88,7 +79,6 @@ public class JwtUtil {
         }
     }
 
-    //사용자 아이디 추출
     public String extractUserId(String token) {
         return Jwts.parserBuilder()
                 .setSigningKey(key)
@@ -98,7 +88,6 @@ public class JwtUtil {
                 .getSubject();
     }
 
-    //커스텀 claim 추출 (userNo, userType 등)
     public Object extractClaim(String token, String claimName) {
         return Jwts.parserBuilder()
                 .setSigningKey(key)
@@ -107,8 +96,7 @@ public class JwtUtil {
                 .getBody()
                 .get(claimName);
     }
-    
-    //쿠키에서 토큰 꺼내기
+
     public String extractTokenFromCookie(HttpServletRequest request, String tokenName) {
         if (request.getCookies() != null) {
             for (Cookie cookie : request.getCookies()) {
